Add files via upload
This Jupyter notebook offers a guide to performing Exploratory Data Analysis (EDA) on an Airbnb dataset. It covers essential tasks such as data loading, initial examination, and data cleaning. The notebook is designed to be accessible and easy to follow, providing a solid foundation for further analysis or machine learning tasks.